requirements.txt
asgiref==3.7.2
Django==4.2.9
django-cors-headers==4.3.1
djangorestframework==3.14.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
xmltodict==0.13.0

pywikibot~=8.6.0
regex~=2023.12.25
wikitextparser~=0.55.8